requirements.txt
python-decouple
tweepy
